Refactor key management - BREAKING API
Phoenix wallet uses `KeyManager.derivePrivateKey()` to extract
key material and uses it directly in its code. This is not compatible
with the idea of segregating the key material behind the specific
`KeyManager` implementation.

We change `LocalKeyManager.derivePrivateKey()` to return a
private key descriptor instead of a direct private key
material. This breaks Phoenix.

Also, we make `derivePrivateKey` part of the `KeyManager`
interface to have something more cohesive. Maybe we don't
want to expose this function to the outside world but if
we do, let's expose it _officially_ through `KeyManager`.

`derivePrivateKey()` is used in two circumstances by Phoenix:
1. for lnUrl related signing ;
2. to generate a symmetric key for cloud storage ciphering.

For 1. `LnurlAuth.signChallenge()` is impacted in the Phoenix
source code and the code should be changed to look something
like that:

```
   fun signChallenge(
            challenge: String,
            key: PrivateKeyDescriptor
        ): Pair<PublicKey, ByteVector> {
            return key.publicKey() to key.sign(ByteVector32.fromValidHex(challenge)
        }
```

For 2. we introduce a new function to the `KeyManager`:
```
deterministicKeyMaterial(keyPath: KeyPath): ByteVector32
```

`LocalKeyManager` implements `deterministicKeyMaterial` in a
way that makes it compatible with the way Phoenix currently
generates the cloud key. Phoenix code in WalletManager.kt
should be changed to look something like that:

```
fun LocalKeyManager.cloudKey(): ByteVector32 {
    val path = KeyPath(if (isMainnet()) "m/51'/0'/0'/0" else "m/51'/1'/0'/0")
    return deterministicKeyMaterial(path)
}
```

Having a way to generate key material from the KeyManager is
very practical indeed as one will not need to save any other seed but
will always be able to retrieve the key and get the data. However
allowing someone to access the key material from random derivation
path of the master key might be problematic or even not possible at
all depending on the crypto engine used, for instance if we use a
secure element or hardware device. With this specification
for `deterministicKeyMaterial` we only ask for implementor to have
a deterministic process given some derivation data that happens to
be a KeyPath but we're not asking for real key derivation from this
path.