Refactor key management - Inline function as it's used only once
`LocalKeyManager.privateKey()` is used once and only once to
generate the `shaSeed` for the `ChannelKeys`. By inlining its
code we make it clear that this function is not supposed to
be used by any other code and that's important as it is a
function that exposes key material to the outside world.

Here we make it clear that it's `LocalKeyManager` accountability
to decide how the `shaSeed` is created so that it can be
deterministically recreated and safely extracted.