Refactor key management - encapsulate Transactions.sign
The signing of transactions uses private key material. We'd rather
encapsulate this behin a `LocalKeyManager`. As a first step, we
move the `sign` function from `Transactions` to a `sign` method
in `PrivateKeyDescriptor`.

We'll have to move things around so that `PrivateKeyDescriptor`
implementation are sorted and places in sensitive places in code.
Right now, it's all over the place but we'll live with that a bit
more.

Pros of making `sign` a method:
* helps encapsulate the actual signing strategy and change it when
  we want to use another `KeyManager`;
* opens the door to limiting what operations could be done with
  which key.

For instance we never `sign` with a
`SwapInOnChainKeys.userPrivateKey` so it might be nice to not be able
to so in the future.