Refactor key management - segregate instantiate() to LocalKeyManager 2
We remove `instantiate()` from `ExtendedPrivateKeyDescriptor` so
that only the `LocalKeyManager` can instantiate actual keys.

To ease the process, we introduce derivation capabilities to
`ExtendedPrivateKeyDescriptor`.

We change `LocalKeyManager` to only rely on this way of
deriving keys. Some part of the code still extract key material
so it may sometimes look a bit clunky with casts to be able
to call the `instantiate()` method.